Reorganized some of the datatypes, modified player.
Limited support for collision detection has been added. The player's LOS is now rendered correctly off of blocks.

Reflections are currently a mess, a lot of cleaning needs to be done before the actual swap is implemented.

Some functions will now generate meaningful errors if the wrong patterns of MObjects are detected.